Extract author related methods into authorable module to share with quizzes
